requirements.txt
aiohttp==3.8.3
aioitertools==0.11.0
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
certifi==2022.9.24
chardet==5.0.0
charset-normalizer==2.1.1
frozenlist==1.3.1
fsspec==2022.10.0
idna==3.4
jmespath==1.0.1
multidict==6.0.2
oauthlib==3.2.2
psycopg2==2.9.4
psycopg2-binary==2.9.4
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
requests==2.28.1
requests-oauthlib==1.3.1
s3fs==2022.10.0
s3transfer==0.6.0
schedule==1.1.0
six==1.16.0
style
tweepy==4.14.0
update==0.0.1
urllib3==1.26.12
wrapt==1.14.1
yarl==1.8.1
